会議を挟んだ際にキルクールが変更されてた場合デフォルトに戻るバグの修正

+++ SuperNewRoles/Helpers/RPCHelper.cs
@@ -153,7 +153,7 @@ public static class RPCHelper
             writer.StartMessage(1); //0x01 Data
             {
                 writer.WritePacked(gm.NetId);
-                writer.StartMessage((byte)4);
+                writer.StartMessage(4);
                 writer.WriteBytesAndSize(gm.LogicOptions.gameOptionsFactory.ToBytes(gameOptions));
                 writer.EndMessage();
             }
